RuboCop Lintern Fix
