+++ .env
@@ -1,2 +1,2 @@
-SLACK_API_TOKEN=xoxb-893526589554-899326518131-4zRNIAVhDpYOwoqaUqcqBwr0
+SLACK_API_TOKEN=xoxb-902877965440-890135764498-mzUnPxuOxu4Sqn50UVC0LTCG
 GIPHY_API_KEY=aNZ45LBFIcpkXMNwmZiAvqq0C8w8FARL
