--- slack-imojionbot.rb
@@ -1,12 +1,0 @@
-require 'slack-ruby-bot'
-require 'slack-imojionbot/commands/cpp'
-require 'slack-imojionbot/commands/css'
-require 'slack-imojionbot/commands/hello'
-require 'slack-imojionbot/commands/html'
-require 'slack-imojionbot/commands/java'
-require 'slack-imojionbot/commands/javascript'
-require 'slack-imojionbot/commands/news'
-require 'slack-imojionbot/commands/python'
-require 'slack-imojionbot/commands/ruby'
-require 'slack-imojionbot/commands/gifme'
-require 'slack-imojionbot/bot'
