--- slack-imojionbot/commands/java.rb
@@ -1,15 +1,0 @@
-require 'rss'
-require 'open-uri'
-
-module SlackImojionbot
-  module Commands
-    class Java < SlackRubyBot::Commands::Base
-        command 'java' do |client, data, _match|
-          url = 'http://blog.jooq.org/feed'
-          rss = RSS::Parser.parse(open(url).read, false).items.first
-          client.say(channel: data.channel, text: "Java and SQL news! fresh from JOOQ Blog")
-          client.say(channel: data.channel, text: rss.link)
-        end
-    end
-  end
-end
