+++ slack_imojionbot.rb
@@ -1,0 +1,12 @@
+require 'slack-ruby-bot'
+require 'slack_imojionbot/commands/cpp'
+require 'slack_imojionbot/commands/css'
+require 'slack_imojionbot/commands/hello'
+require 'slack_imojionbot/commands/html'
+require 'slack_imojionbot/commands/java'
+require 'slack_imojionbot/commands/javascript'
+require 'slack_imojionbot/commands/news'
+require 'slack_imojionbot/commands/python'
+require 'slack_imojionbot/commands/ruby'
+require 'slack_imojionbot/commands/gifme'
+require 'slack_imojionbot/bot'
