+++ slack_imojionbot/commands/cpp.rb
@@ -1,0 +1,15 @@
+require 'rss'
+require 'open-uri'
+
+module SlackImojionbot
+  module Commands
+    class Cpp < SlackRubyBot::Commands::Base
+      command 'cpp' do |client, data, _match|
+        url = 'https://blogs.msdn.microsoft.com/vcblog/feed/'
+        rss = RSS::Parser.parse(open(url).read, false).items.first
+        client.say(channel: data.channel, text: 'c++ news for you!, here you go!')
+        client.say(channel: data.channel, text: rss.link)
+      end
+    end
+  end
+end
