+++ slack_imojionbot/commands/gifme.rb
@@ -1,10 +1,9 @@
-
 module SlackGoogleBot
   module Commands
     class Gifme < SlackRubyBot::Commands::Base
       match(/^(?<bot>\w*)\s(?<gifme>\w+)\s(?<expression>.*)$/) do |client, data, match|
         expression = match['expression'].strip
-        client.say(text: "Gifyou!", channel: data.channel, gif: expression)
+        client.say(text: 'Gifyou!', channel: data.channel, gif: expression)
       end
     end
   end
