+++ spec/slack_imojionbot/bot_spec.rb
