+++ spec/slack_imojionbot/commands/html_spec.rb
