+++ spec/slack_imojionbot/commands/java_spec.rb
