+++ spec/slack_imojionbot/commands/javascript_spec.rb
