+++ spec/slack_imojionbot/commands/news_spec.rb
