+++ spec/slack_imojionbot/commands/python_spec.rb
