+++ spec/slack_imojionbot/commands/ruby_spec.rb
