+++ spec/spec_helper.rb
@@ -1,4 +1,4 @@
 $LOAD_PATH.unshift(File.join(File.dirname(__FILE__), '..'))
 
 require 'slack-ruby-bot/rspec'
-require 'slack-imojionbot'
+require 'slack_imojionbot'
