+++ vcr_setup.rb
@@ -4,5 +4,5 @@ VCR.config do |c|
   c.allow_http_connections_when_no_cassette = true
   c.cassette_library_dir = 'vcr_cassettes'
   c.stub_with :webmock
-  c.default_cassette_options = { :record => :once }
+  c.default_cassette_options = { key: 'once' }
 end
